+++ src/Data/Morpheus/Error/Fragment.hs
@@ -1,15 +1,22 @@
+{-# LANGUAGE NamedFieldPuns    #-}
 {-# LANGUAGE OverloadedStrings #-}
 
 module Data.Morpheus.Error.Fragment
   ( cannotSpreadWithinItself
+  , unusedFragment
+  , unknownFragment
+  , cannotBeSpreadOnType
+  , fragmentNameCollision
   ) where
 
--- import Data.Morpheus.Error.Utils (errorMessage)
-import           Data.Morpheus.Types.Internal.Base       (EnhancedKey (..))
-import           Data.Morpheus.Types.Internal.Validation (GQLError (..), GQLErrors)
-
--- import Data.Morpheus.Types.Internal.Base (MetaInfo(..))
+import           Data.Semigroup                          ((<>))
+import           Data.Text                               (Text)
 import qualified Data.Text                               as T
+
+-- MORPHEUS
+import           Data.Morpheus.Error.Utils               (errorMessage)
+import           Data.Morpheus.Types.Internal.Base       (EnhancedKey (..), Position)
+import           Data.Morpheus.Types.Internal.Validation (GQLError (..), GQLErrors)
 
 {-
   FRAGMENT:
@@ -22,6 +29,18 @@ import qualified Data.Text                               as T
     fragment H on D {...}  ->  "Unknown type \"D\"."
     {...H} -> "Unknown fragment \"H\"."
 -}
+fragmentNameCollision :: [EnhancedKey] -> GQLErrors
+fragmentNameCollision = map toError
+  where
+    toError EnhancedKey {uid, location} =
+      GQLError {desc = "There can be only one fragment named \"" <> uid <> "\".", positions = [location]}
+
+unusedFragment :: [EnhancedKey] -> GQLErrors
+unusedFragment = map toError
+  where
+    toError EnhancedKey {uid, location} =
+      GQLError {desc = "Fragment \"" <> uid <> "\" is never used.", positions = [location]}
+
 cannotSpreadWithinItself :: [EnhancedKey] -> GQLErrors
 cannotSpreadWithinItself fragments = [GQLError {desc = text, positions = map location fragments}]
   where
@@ -33,3 +52,26 @@ cannotSpreadWithinItself fragments = [GQLError {desc = text, positions = map loc
         , T.intercalate "," (map uid fragments)
         , "."
         ]
+
+-- {...H} -> "Unknown fragment \"H\"."
+unknownFragment :: Text -> Position -> GQLErrors
+unknownFragment key' position' = errorMessage position' text
+  where
+    text = T.concat ["Unknown Fragment \"", key', "\"."]
+
+-- Fragment type mismatch -> "Fragment \"H\" cannot be spread here as objects of type \"Hobby\" can never be of type \"Experience\"."
+cannotBeSpreadOnType :: Maybe Text -> Text -> Position -> Text -> GQLErrors
+cannotBeSpreadOnType key' type' position' selectionType' = errorMessage position' text
+  where
+    text =
+      T.concat
+        [ "Fragment"
+        , getName key'
+        , " cannot be spread here as objects of type \""
+        , selectionType'
+        , "\" can never be of type \""
+        , type'
+        , "\"."
+        ]
+    getName (Just x') = T.concat [" \"", x', "\""]
+    getName Nothing   = ""