+++ src/Data/Morpheus/Parser/Parser.hs
@@ -6,7 +6,7 @@ module Data.Morpheus.Parser.Parser
   ) where
 
 import qualified Data.List.NonEmpty                      as NonEmpty
-import           Data.Map                                (fromList, toList)
+import           Data.Map                                (toList)
 import           Data.Maybe                              (maybe)
 import           Data.Morpheus.Parser.Fragment           (fragment)
 import           Data.Morpheus.Parser.Internal           (Parser)
@@ -27,9 +27,9 @@ request :: Parser GQLQueryRoot
 request =
   label "GQLQueryRoot" $ do
     space
-    operator' <- parseAnonymousQuery <|> parseOperator
-    fragmentLib <- fromList <$> manyTill fragment eof
-    pure GQLQueryRoot {queryBody = operator', fragments = fragmentLib, inputVariables = []}
+    operator <- parseAnonymousQuery <|> parseOperator
+    fragments <- manyTill fragment eof
+    pure GQLQueryRoot {operator, fragments, inputVariables = []}
 
 processErrorBundle :: ParseErrorBundle Text Void -> GQLErrors
 processErrorBundle = fmap parseErrorToGQLError . bundleToErrors