+++ src/Data/Morpheus/Validation/Selection.hs
@@ -19,16 +19,16 @@ import           Data.Morpheus.Types.Internal.Data             (DataField (..), 
                                                                 DataTypeKind (..), DataTypeLib (..))
 import           Data.Morpheus.Types.Internal.Validation       (Validation)
 import           Data.Morpheus.Validation.Arguments            (validateArguments)
-import           Data.Morpheus.Validation.Spread               (castFragmentType, resolveSpread)
+import           Data.Morpheus.Validation.Fragment             (castFragmentType, resolveSpread)
 import           Data.Morpheus.Validation.Utils.Selection      (lookupFieldAsSelectionSet, lookupSelectionField,
                                                                 lookupUnionTypes, notObject)
 import           Data.Morpheus.Validation.Utils.Utils          (checkNameCollision)
 import           Data.Text                                     (Text)
 
 checkDuplicatesOn :: DataOutputObject -> SelectionSet -> Validation SelectionSet
-checkDuplicatesOn DataType {typeName = name'} keys = checkNameCollision enhancedKeys (map fst keys) error' >> pure keys
+checkDuplicatesOn DataType {typeName = name'} keys = checkNameCollision enhancedKeys selError >> pure keys
   where
-    error' = duplicateQuerySelections name'
+    selError = duplicateQuerySelections name'
     enhancedKeys = map selToKey keys
     selToKey (key', Selection {selectionPosition = position'}) = EnhancedKey key' position'